+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 _release-date_
 
+* **Breaking:** Calling `setVisibilityHint(false)` on a LithoView will ignore mounting calls until `setVisibilityHint(true)` is called.
+
 * TBA
 
 For more details, see the [full diff](https://github.com/facebook/litho/compare/v0.39.0...master).